fix(saves): reputation 改 float64
跟 morale/fatigue/loyalty 同根 — 遊戲邏輯對 reputation 做小數運算
（事件 delta 經 multiplier 後可能落在 .5），前端傳 96.5 在 json.Unmarshal
階段失敗，回 4001 (CodeBadPayload)。

- Reputation int → float64
- sanityCheck %d → %g
- spec doc int → number，標註可為小數

+++ internal/save/save_handler.go
@@ -71,7 +71,7 @@ type ConflictData struct {
 type saveDataForCheck struct {
 	Day               int          `json:"day"`
 	Money             int          `json:"money"`
-	Reputation        int          `json:"reputation"`
+	Reputation        float64      `json:"reputation"`
 	TierBudget        int          `json:"tierBudget"`
 	OfficeLevel       int          `json:"officeLevel"`
 	CompanyBuffs      companyBuffs `json:"companyBuffs"`
@@ -248,7 +248,7 @@ func sanityCheck(d *saveDataForCheck) error {
 		return fmt.Errorf("money must be >= 0 (got %d)", d.Money)
 	}
 	if d.Reputation < 0 || d.Reputation > 100 {
-		return fmt.Errorf("reputation must be in [0,100] (got %d)", d.Reputation)
+		return fmt.Errorf("reputation must be in [0,100] (got %g)", d.Reputation)
 	}
 	if d.TierBudget < 0 {
 		return fmt.Errorf("tierBudget must be >= 0 (got %d)", d.TierBudget)
